+++ android/versioned-abis/expoview-abi42_0_0/src/main/java/abi42_0_0/expo/modules/av/player/SimpleExoPlayerData.java
@@ -5,28 +5,25 @@ import android.net.Uri;
 import android.os.Bundle;
 import android.os.Handler;
 
-import androidx.annotation.Nullable;
-
 import android.text.TextUtils;
 import android.util.Log;
 import android.util.Pair;
 import android.view.Surface;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
 import com.google.android.exoplayer2.C;
-import com.google.android.exoplayer2.DefaultLoadControl;
-import com.google.android.exoplayer2.DefaultRenderersFactory;
 import com.google.android.exoplayer2.ExoPlaybackException;
-import com.google.android.exoplayer2.ExoPlayerFactory;
-import com.google.android.exoplayer2.Format;
+import com.google.android.exoplayer2.MediaItem;
 import com.google.android.exoplayer2.PlaybackParameters;
 import com.google.android.exoplayer2.Player;
 import com.google.android.exoplayer2.SimpleExoPlayer;
-import com.google.android.exoplayer2.Timeline;
-import com.google.android.exoplayer2.extractor.DefaultExtractorsFactory;
-import com.google.android.exoplayer2.source.AdaptiveMediaSourceEventListener;
 import com.google.android.exoplayer2.source.ExtractorMediaSource;
+import com.google.android.exoplayer2.source.LoadEventInfo;
+import com.google.android.exoplayer2.source.MediaLoadData;
 import com.google.android.exoplayer2.source.MediaSource;
-import com.google.android.exoplayer2.source.TrackGroupArray;
+import com.google.android.exoplayer2.source.MediaSourceEventListener;
 import com.google.android.exoplayer2.source.dash.DashMediaSource;
 import com.google.android.exoplayer2.source.dash.DefaultDashChunkSource;
 import com.google.android.exoplayer2.source.hls.HlsMediaSource;
@@ -34,8 +31,6 @@ import com.google.android.exoplayer2.source.smoothstreaming.DefaultSsChunkSource
 import com.google.android.exoplayer2.source.smoothstreaming.SsMediaSource;
 import com.google.android.exoplayer2.trackselection.AdaptiveTrackSelection;
 import com.google.android.exoplayer2.trackselection.DefaultTrackSelector;
-import com.google.android.exoplayer2.trackselection.TrackSelection;
-import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
 import com.google.android.exoplayer2.trackselection.TrackSelector;
 import com.google.android.exoplayer2.upstream.BandwidthMeter;
 import com.google.android.exoplayer2.upstream.DataSource;
@@ -43,32 +38,33 @@ import com.google.android.exoplayer2.upstream.DataSpec;
 import com.google.android.exoplayer2.upstream.DefaultBandwidthMeter;
 import com.google.android.exoplayer2.upstream.RawResourceDataSource;
 import com.google.android.exoplayer2.util.Util;
+import com.google.android.exoplayer2.video.VideoListener;
 
 import java.io.IOException;
 import java.util.Map;
 
 import abi42_0_0.expo.modules.av.AVManagerInterface;
 import abi42_0_0.expo.modules.av.AudioFocusNotAcquiredException;
-import abi42_0_0.expo.modules.av.player.datasource.CustomHeadersOkHttpDataSourceFactory;
 import abi42_0_0.expo.modules.av.player.datasource.DataSourceFactoryProvider;
 
 class SimpleExoPlayerData extends PlayerData
-  implements Player.EventListener, ExtractorMediaSource.EventListener, SimpleExoPlayer.VideoListener, AdaptiveMediaSourceEventListener {
+    implements Player.EventListener, MediaSourceEventListener, VideoListener {
 
   private static final String IMPLEMENTATION_NAME = "SimpleExoPlayer";
   private static final String TAG = SimpleExoPlayerData.class.getSimpleName();
 
   private SimpleExoPlayer mSimpleExoPlayer = null;
-  private String mOverridingExtension;
+  private final String mOverridingExtension;
   private LoadCompletionListener mLoadCompletionListener = null;
   private boolean mFirstFrameRendered = false;
   private Pair<Integer, Integer> mVideoWidthHeight = null;
   private Integer mLastPlaybackState = null;
   private boolean mIsLooping = false;
   private boolean mIsLoading = true;
-  private Context mReactContext;
+  private final Context mReactContext;
 
-  SimpleExoPlayerData(final AVManagerInterface avModule, final Context context, final Uri uri, final String overridingExtension, final Map<String, Object> requestHeaders) {
+  SimpleExoPlayerData(final AVManagerInterface avModule, final Context context, final Uri uri,
+      final String overridingExtension, final Map<String, Object> requestHeaders) {
     super(avModule, uri, requestHeaders);
     mReactContext = context;
     mOverridingExtension = overridingExtension;
@@ -87,29 +83,31 @@ class SimpleExoPlayerData extends PlayerData
   public void load(final Bundle status, final LoadCompletionListener loadCompletionListener) {
     mLoadCompletionListener = loadCompletionListener;
 
-    // Create a default TrackSelector
-    final Handler mainHandler = new Handler();
-    // Measures bandwidth during playback. Can be null if not required.
-    final BandwidthMeter bandwidthMeter = new DefaultBandwidthMeter();
-    final TrackSelection.Factory trackSelectionFactory = new AdaptiveTrackSelection.Factory();
-    final TrackSelector trackSelector = new DefaultTrackSelector(trackSelectionFactory);
+    final Context context = mAVModule.getContext();
+    final BandwidthMeter bandwidthMeter = new DefaultBandwidthMeter.Builder(context).build();
+    final TrackSelector trackSelector = new DefaultTrackSelector(context, new AdaptiveTrackSelection.Factory());
 
     // Create the player
-    mSimpleExoPlayer = ExoPlayerFactory.newSimpleInstance(
-      mAVModule.getContext(),
-      new DefaultRenderersFactory(mAVModule.getContext()),
-      trackSelector,
-      new DefaultLoadControl(),
-      null,
-      bandwidthMeter);
+    mSimpleExoPlayer = new SimpleExoPlayer.Builder(context)
+        .setTrackSelector(trackSelector)
+        .setBandwidthMeter(bandwidthMeter)
+        .build();
+
     mSimpleExoPlayer.addListener(this);
     mSimpleExoPlayer.addVideoListener(this);
 
     // Produces DataSource instances through which media data is loaded.
-    final DataSource.Factory dataSourceFactory = mAVModule.getModuleRegistry().getModule(DataSourceFactoryProvider.class).createFactory(mReactContext, mAVModule.getModuleRegistry(), Util.getUserAgent(mAVModule.getContext(), "yourApplicationName"), mRequestHeaders, bandwidthMeter.getTransferListener());
+    final DataSource.Factory dataSourceFactory = mAVModule.getModuleRegistry()
+        .getModule(DataSourceFactoryProvider.class)
+        .createFactory(
+            mReactContext,
+            mAVModule.getModuleRegistry(),
+            Util.getUserAgent(context, "yourApplicationName"),
+            mRequestHeaders,
+            bandwidthMeter.getTransferListener());
     try {
       // This is the MediaSource representing the media to be played.
-      final MediaSource source = buildMediaSource(mUri, mOverridingExtension, mainHandler, dataSourceFactory);
+      final MediaSource source = buildMediaSource(mUri, mOverridingExtension, dataSourceFactory);
 
       // Prepare the player with the source.
       mSimpleExoPlayer.prepare(source);
@@ -121,6 +119,7 @@ class SimpleExoPlayerData extends PlayerData
 
   @Override
   public synchronized void release() {
+    stopUpdatingProgressIfNecessary();
     if (mSimpleExoPlayer != null) {
       mSimpleExoPlayer.release();
       mSimpleExoPlayer = null;
@@ -153,7 +152,7 @@ class SimpleExoPlayerData extends PlayerData
 
   @Override
   void applyNewStatus(final Integer newPositionMillis, final Boolean newIsLooping)
-    throws AudioFocusNotAcquiredException, IllegalStateException {
+      throws AudioFocusNotAcquiredException, IllegalStateException {
     if (mSimpleExoPlayer == null) {
       throw new IllegalStateException("mSimpleExoPlayer is null!");
     }
@@ -200,14 +199,14 @@ class SimpleExoPlayerData extends PlayerData
     final int duration = (int) mSimpleExoPlayer.getDuration();
     map.putInt(STATUS_DURATION_MILLIS_KEY_PATH, duration);
     map.putInt(STATUS_POSITION_MILLIS_KEY_PATH,
-      getClippedIntegerForValue((int) mSimpleExoPlayer.getCurrentPosition(), 0, duration));
+        getClippedIntegerForValue((int) mSimpleExoPlayer.getCurrentPosition(), 0, duration));
     map.putInt(STATUS_PLAYABLE_DURATION_MILLIS_KEY_PATH,
-      getClippedIntegerForValue((int) mSimpleExoPlayer.getBufferedPosition(), 0, duration));
+        getClippedIntegerForValue((int) mSimpleExoPlayer.getBufferedPosition(), 0, duration));
 
     map.putBoolean(STATUS_IS_PLAYING_KEY_PATH,
-      mSimpleExoPlayer.getPlayWhenReady() && mSimpleExoPlayer.getPlaybackState() == Player.STATE_READY);
+        mSimpleExoPlayer.getPlayWhenReady() && mSimpleExoPlayer.getPlaybackState() == Player.STATE_READY);
     map.putBoolean(STATUS_IS_BUFFERING_KEY_PATH,
-      mIsLoading || mSimpleExoPlayer.getPlaybackState() == Player.STATE_BUFFERING);
+        mIsLoading || mSimpleExoPlayer.getPlaybackState() == Player.STATE_BUFFERING);
 
     map.putBoolean(STATUS_IS_LOOPING_KEY_PATH, mIsLooping);
   }
@@ -233,7 +232,7 @@ class SimpleExoPlayerData extends PlayerData
 
   // --------- Interface implementation ---------
 
-  // AudioEventHandler
+  // region AudioEventHandler
 
   @Override
   public void pauseImmediately() {
@@ -255,7 +254,9 @@ class SimpleExoPlayerData extends PlayerData
     }
   }
 
-  // ExoPlayer.EventListener
+  // endregion
+
+  // region ExoPlayer.EventListener
 
   @Override
   public void onLoadingChanged(final boolean isLoading) {
@@ -282,12 +283,6 @@ class SimpleExoPlayerData extends PlayerData
   }
 
   @Override
-  public void onTracksChanged(TrackGroupArray trackGroups,
-                              TrackSelectionArray trackSelections) {
-
-  }
-
-  @Override
   public void onPlayerStateChanged(final boolean playWhenReady, final int playbackState) {
     if (playbackState == Player.STATE_READY && mLoadCompletionListener != null) {
       final LoadCompletionListener listener = mLoadCompletionListener;
@@ -296,8 +291,8 @@ class SimpleExoPlayerData extends PlayerData
     }
 
     if (mLastPlaybackState != null
-      && playbackState != mLastPlaybackState
-      && playbackState == Player.STATE_ENDED) {
+        && playbackState != mLastPlaybackState
+        && playbackState == Player.STATE_ENDED) {
       callStatusUpdateListenerWithDidJustFinish();
       stopUpdatingProgressIfNecessary();
     } else {
@@ -320,7 +315,8 @@ class SimpleExoPlayerData extends PlayerData
     // > A period defines a single logical piece of media, for example a media file.
     // > It may also define groups of ads inserted into the media,
     // > along with information about whether those ads have been loaded and played.
-    // Source: https://google.github.io/ExoPlayer/doc/reference/com/google/android/exoplayer2/Timeline.Period.html
+    // Source:
+    // https://google.github.io/ExoPlayer/doc/reference/com/google/android/exoplayer2/Timeline.Period.html
     // So I guess it's safe to say that when a period transition happens,
     // media file transition happens, so we just finished playing one.
     if (reason == Player.DISCONTINUITY_REASON_PERIOD_TRANSITION) {
@@ -328,11 +324,18 @@ class SimpleExoPlayerData extends PlayerData
     }
   }
 
+  // endregion
 
-  // ExtractorMediaSource.EventListener
+  // region MediaSourceEventListener
 
   @Override
-  public void onLoadError(final IOException error) {
+  public void onLoadError(
+      int windowIndex,
+      @Nullable MediaSource.MediaPeriodId mediaPeriodId,
+      LoadEventInfo loadEventInfo,
+      MediaLoadData mediaLoadData,
+      IOException error,
+      boolean wasCanceled) {
     if (mLoadCompletionListener != null) {
       final LoadCompletionListener listener = mLoadCompletionListener;
       mLoadCompletionListener = null;
@@ -351,10 +354,13 @@ class SimpleExoPlayerData extends PlayerData
     release();
   }
 
-  // SimpleExoPlayer.VideoListener
+  // endregion
+
+  // region VideoListener
 
   @Override
-  public void onVideoSizeChanged(final int width, final int height, final int unAppliedRotationDegrees, final float pixelWidthHeightRatio) {
+  public void onVideoSizeChanged(final int width, final int height, final int unAppliedRotationDegrees,
+      final float pixelWidthHeightRatio) {
     // TODO other params?
     mVideoWidthHeight = new Pair<>(width, height);
     if (mFirstFrameRendered && mVideoSizeUpdateListener != null) {
@@ -370,11 +376,13 @@ class SimpleExoPlayerData extends PlayerData
     mFirstFrameRendered = true;
   }
 
-  // https://github.com/google/ExoPlayer/blob/2b20780482a9c6b07416bcbf4de829532859d10a/demos/main/src/main/java/com/google/android/exoplayer2/demo/PlayerActivity.java#L365-L393
-  private MediaSource buildMediaSource(Uri uri, String overrideExtension, Handler mainHandler, DataSource.Factory factory) {
+  // endregion
+
+  private MediaSource buildMediaSource(@NonNull Uri uri, String overrideExtension, DataSource.Factory factory) {
     try {
       if (uri.getScheme() == null) {
-        int resourceId = mReactContext.getResources().getIdentifier(uri.toString(), "raw", mReactContext.getPackageName());
+        int resourceId = mReactContext.getResources().getIdentifier(uri.toString(), "raw",
+            mReactContext.getPackageName());
         DataSpec dataSpec = new DataSpec(RawResourceDataSource.buildRawResourceUri(resourceId));
         final RawResourceDataSource rawResourceDataSource = new RawResourceDataSource(mReactContext);
         rawResourceDataSource.open(dataSpec);
@@ -383,67 +391,57 @@ class SimpleExoPlayerData extends PlayerData
     } catch (Exception e) {
       Log.e(TAG, "Error reading raw resource from ExoPlayer", e);
     }
-    @C.ContentType int type = TextUtils.isEmpty(overrideExtension) ? Util.inferContentType(String.valueOf(uri)) : Util.inferContentType("." + overrideExtension);
+    @C.ContentType
+    int type = TextUtils.isEmpty(overrideExtension) ? Util.inferContentType(String.valueOf(uri))
+        : Util.inferContentType("." + overrideExtension);
     switch (type) {
       case C.TYPE_SS:
-        return new SsMediaSource(uri, factory,
-          new DefaultSsChunkSource.Factory(factory), mainHandler, this);
+        return new SsMediaSource.Factory(new DefaultSsChunkSource.Factory(factory), factory)
+            .createMediaSource(MediaItem.fromUri(uri));
       case C.TYPE_DASH:
-        return new DashMediaSource(uri, factory,
-          new DefaultDashChunkSource.Factory(factory), mainHandler, this);
+        return new DashMediaSource.Factory(new DefaultDashChunkSource.Factory(factory), factory)
+            .createMediaSource(MediaItem.fromUri(uri));
       case C.TYPE_HLS:
-        return new HlsMediaSource(uri, factory, mainHandler, this);
+        return new HlsMediaSource.Factory(factory).createMediaSource(MediaItem.fromUri(uri));
       case C.TYPE_OTHER:
-        return new ExtractorMediaSource(uri, factory, new DefaultExtractorsFactory(), mainHandler, this);
+        return new ExtractorMediaSource.Factory(factory).createMediaSource(MediaItem.fromUri(uri));
       default: {
         throw new IllegalStateException("Content of this type is unsupported at the moment. Unsupported type: " + type);
       }
     }
   }
 
-  // AdaptiveMediaSourceEventListener
+  // region MediaSourceEventListener
+
   @Override
-  public void onMediaPeriodCreated(int windowIndex, MediaSource.MediaPeriodId mediaPeriodId) {
+  public void onLoadStarted(int windowIndex, @Nullable MediaSource.MediaPeriodId mediaPeriodId,
+      LoadEventInfo loadEventInfo, MediaLoadData mediaLoadData) {
 
   }
 
   @Override
-  public void onMediaPeriodReleased(int windowIndex, MediaSource.MediaPeriodId mediaPeriodId) {
+  public void onLoadCompleted(int windowIndex, @Nullable MediaSource.MediaPeriodId mediaPeriodId,
+      LoadEventInfo loadEventInfo, MediaLoadData mediaLoadData) {
 
   }
 
   @Override
-  public void onLoadStarted(int windowIndex, @Nullable MediaSource.MediaPeriodId mediaPeriodId, LoadEventInfo loadEventInfo, MediaLoadData mediaLoadData) {
+  public void onLoadCanceled(int windowIndex, @Nullable MediaSource.MediaPeriodId mediaPeriodId,
+      LoadEventInfo loadEventInfo, MediaLoadData mediaLoadData) {
 
   }
 
   @Override
-  public void onLoadCompleted(int windowIndex, @Nullable MediaSource.MediaPeriodId mediaPeriodId, LoadEventInfo loadEventInfo, MediaLoadData mediaLoadData) {
+  public void onUpstreamDiscarded(int windowIndex, MediaSource.MediaPeriodId mediaPeriodId,
+      MediaLoadData mediaLoadData) {
 
   }
 
   @Override
-  public void onLoadCanceled(int windowIndex, @Nullable MediaSource.MediaPeriodId mediaPeriodId, LoadEventInfo loadEventInfo, MediaLoadData mediaLoadData) {
+  public void onDownstreamFormatChanged(int windowIndex, @Nullable MediaSource.MediaPeriodId mediaPeriodId,
+      MediaLoadData mediaLoadData) {
 
   }
 
-  @Override
-  public void onLoadError(int windowIndex, @Nullable MediaSource.MediaPeriodId mediaPeriodId, LoadEventInfo loadEventInfo, MediaLoadData mediaLoadData, IOException error, boolean wasCanceled) {
-
-  }
-
-  @Override
-  public void onReadingStarted(int windowIndex, MediaSource.MediaPeriodId mediaPeriodId) {
-
-  }
-
-  @Override
-  public void onUpstreamDiscarded(int windowIndex, MediaSource.MediaPeriodId mediaPeriodId, MediaLoadData mediaLoadData) {
-
-  }
-
-  @Override
-  public void onDownstreamFormatChanged(int windowIndex, @Nullable MediaSource.MediaPeriodId mediaPeriodId, MediaLoadData mediaLoadData) {
-
-  }
+  // endregion
 }