+++ android/versioned-abis/expoview-abi44_0_0/src/main/java/abi44_0_0/expo/modules/av/video/VideoViewManager.java
@@ -2,11 +2,10 @@ package abi44_0_0.expo.modules.av.video;
 
 import android.content.Context;
 
-import com.yqritc.scalablevideoview.ScalableType;
-
 import java.util.ArrayList;
 import java.util.List;
 
+import abi44_0_0.expo.modules.av.video.scalablevideoview.ScalableType;
 import abi44_0_0.expo.modules.core.ModuleRegistry;
 import abi44_0_0.expo.modules.core.ViewManager;
 import abi44_0_0.expo.modules.core.arguments.ReadableArguments;