+++ packages/expo-av/android/src/main/java/expo/modules/av/video/VideoViewManager.java
@@ -2,11 +2,10 @@ package expo.modules.av.video;
 
 import android.content.Context;
 
-import com.yqritc.scalablevideoview.ScalableType;
-
 import java.util.ArrayList;
 import java.util.List;
 
+import expo.modules.av.video.scalablevideoview.ScalableType;
 import expo.modules.core.ModuleRegistry;
 import expo.modules.core.ViewManager;
 import expo.modules.core.arguments.ReadableArguments;